+++ test/matchers/validations_test.rb
@@ -7,9 +7,6 @@ describe "Validations" do
     it { must validate_presence_of(:name) }
     it { must validate_presence_of(:role).with_message("no role") }
 
-    it { must validate_confirmation_of(:password) }
-    it { must validate_acceptance_of(:terms_of_use) }
-
     it { must validate_uniqueness_of(:login).case_insensitive }
     it { must validate_uniqueness_of(:login).scoped_to(:site) }
 
@@ -27,5 +24,8 @@ describe "Validations" do
 
     it { must validate_inclusion_of(:role).to_allow("user", "admin") }
     it { must validate_exclusion_of(:email).to_not_allow("foo@bar.com", "fizz@buzz.com") }
+
+    it { must validate_confirmation_of(:password) }
+    it { must validate_acceptance_of(:terms_of_use).accept_with("1") }
   end
 end